+++ lib/cfilters.h
@@ -359,16 +359,6 @@ bool Curl_conn_cf_wants_httpsrr(struct Curl_cfilter *cf,
 #define CURL_CF_SSL_ENABLE   1
 
 /**
- * Bring the filter chain at `sockindex` for connection `data->conn` into
- * connected state. Which will set `*done` to TRUE.
- * This can be called on an already connected chain with no side effects.
- * When not `blocking`, calls may return without error and `*done != TRUE`,
- * while the individual filters negotiated the connection.
- */
-CURLcode Curl_conn_connect(struct Curl_easy *data, int sockindex,
-                           bool blocking, bool *done);
-
-/**
  * Check if a filter chain at `sockindex` for connection `conn` exists.
  */
 bool Curl_conn_is_setup(struct connectdata *conn, int sockindex);
@@ -430,6 +420,12 @@ unsigned char Curl_conn_get_transport(struct Curl_easy *data,
 /* Get the negotiated ALPN protocol or NULL if none in play */
 const char *Curl_conn_get_alpn_negotiated(struct Curl_easy *data,
                                           struct connectdata *conn);
+
+void Curl_conn_cntrl_update_info(struct Curl_easy *data,
+                                 struct connectdata *conn);
+
+void Curl_conn_remove_setup_filters(struct Curl_easy *data,
+                                    int sockindex);
 
 /**
  * Shutdown the connection at `sockindex` non-blocking, using timeout
